refactor(decision-analyzer/plots)!: split plots.py into a package
Convert the 1698-line python/pdstools/decision_analyzer/plots.py into a
plots/ package, grouped by concern. Method implementations live in
underscore-prefixed private submodules and are bound onto the public
Plot class in __init__.py. Free helper functions (offer-quality, win
distribution, parameter boxplots, component overview) are re-exported
unchanged. The public dir() surface is identical to the previous
single-file module.

Submodules:
  _common            shared constants / boxplot point cap
  _sensitivity       sensitivity, threshold deciles, prio-factor boxplots
  _winloss           global win/loss + win-distribution bar chart
  _optionality       propensity-vs-optionality, optionality per stage / trend
  _funnel            optionality funnel, decision funnel, decisions-without-actions
  _distribution      treemap, action variation, histograms, rank/parameter boxplots
  _components        filter-component plots and prioritization-component distributions
  _trend             generic trend chart
  _offer_quality     offer-quality pie charts and trend (free functions)

This PR also performs the deferred snake_case rename of the
`propensityTH` / `priorityTH` keyword arguments to `propensity_th` /
`priority_th`. The previous v5 PR (v5-da-gold-standard) explicitly
deferred this change because it touches plots.py and many call sites;
since this PR is restructuring plots.py we take the opportunity to do
the rename in one shot. Affected callables:

  - DecisionAnalyzer.get_offer_quality(propensity_th=..., priority_th=...)
  - plots.offer_quality_piecharts(propensity_th=...)
  - plots.offer_quality_single_pie(propensity_th=...)

All call sites in pdstools, the Streamlit apps, and the test suite have
been updated. CHANGELOG.md and docs/migration-v4-to-v5.md document the
rename for v5 users.
